Add support for property initializer syntax
This is an experimental feature:
  https://babeljs.io/docs/plugins/transform-class-properties/

However, example code from the react docs contains the following example:

```jsx
class LoggingButton extends React.Component {
  // This syntax ensures `this` is bound within handleClick.
  // Warning: this is *experimental* syntax.
  handleClick = () => {
    console.log('this is:', this);
  }

  render() {
    return (
      <button onClick={this.handleClick}>
        Click me
      </button>
    );
  }
}
```

Since our projects probably all use babel with support for property initializer
syntax, I think it makes sense to use an eslint rule that also supports them.

Maybe we do not want to support "experimental" features to avoid issues down the
road, but given that this is the "react way" we would end up changing this in
all our react projects anyway, so doing it here means less duplication across
projects. I added it to core config since this is not really a react specific
feature and is valid in any project using babel.